requirements.txt
langchain-google-genai
python-dotenv
wikipedia-api
textblob
deep-translator
langdetect
requests
streamlit
sqlite3
sqlalchemy